+++ src/components/pages/variable/relevant-articles/slider/index.js
@@ -1,0 +1,3 @@
+import Slider from './slider';
+
+export default Slider;
